Split gsettings schemas into a -common subpackage
Similar to how mutter splits its schemas into mutter-common.

This will allow Phosh to depend on a smaller set of packages, since it
currently pulls in the entirety of GNOME Shell in order to pull in
gnome-shell's gschema.

Debian already does this with a package named the same
(gnome-shell-common), as does Alpine (gnome-shell-schemas).

Signed-off-by: Sam Day <me@samcday.com>