Remove keyboard status patches and reapply favorite apps patch
The keyboard status patches are needed for live user mode, but that is
not available yet in Fedora 42. There's no point in carrying these
patches downstream for years and then forgetting to revert them every
release. We just need to land them upstream.

Also, restore Firefox to the favorite apps. Looks like this was dropped
by mistake.